+++ src/Components/Kebab.tsx
@@ -1,6 +1,5 @@
-import "../styles/CardKebab.css";
 import useModalOpen from "../Hooks/useModalOpen";
-import ModalAddLinkToFolder from "./Modal/ModalAddLinkToFolder";
+import ModalAddLink from "./Modal/ModalAddLink";
 import ModalDeleteLink from "./Modal/ModalDeleteLink";
 
 interface Props {
@@ -18,7 +17,7 @@ export function Kebab({ url }: Props) {
         폴더에 추가
       </button>
       {isOpen && clickValue === "폴더에 추가" ? (
-        <ModalAddLinkToFolder
+        <ModalAddLink
           isOpen={isOpen}
           onClose={() => setIsOpen(false)}
           selectedFolderName={"내용 입력"}
